feat: macOS native ビルド対応と Delegate のプラットフォーム分離
- UNUserNotificationCenterDelegate 本体を NotificationHandler に集約
  （全プラットフォーム共通シグネチャを活かした1クラス化）
- AppDelegate (iOS/visionOS) / MacAppDelegate (macOS) に分離し、
  IntentTodoApp で UIApplicationDelegateAdaptor / NSApplicationDelegateAdaptor を #if 分岐
- macOS native で使えない UIKit API (.topBarTrailing, .navigationBarTitleDisplayMode,
  @UIApplicationDelegateAdaptor) を os(iOS) || os(visionOS) ガードに統一
- project.pbxproj: Watch/LiveActivity の Embed を platformFilter = ios; で除外
